--- src/main/java/com/example/newspeed/dto/UsersResponseDto.java
@@ -1,6 +1,0 @@
-package com.example.newspeed.dto;
-
-public class UsersResponseDto {
-
-
-}
