+++ src/main/java/com/example/newspeed/dto/user/LoginUserRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.newspeed.dto;
+package com.example.newspeed.dto.user;
 
 import lombok.Getter;
