+++ src/main/java/com/example/newspeed/dto/user/SignupUserResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.newspeed.dto;
+package com.example.newspeed.dto.user;
 
 import com.example.newspeed.entity.Users;
 import lombok.Getter;
