+++ src/main/java/com/example/newspeed/dto/user/UpdateImageRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.newspeed.dto;
+package com.example.newspeed.dto.user;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
