+++ src/main/java/com/example/newspeed/dto/user/UsersResponseDto.java
@@ -1,0 +1,6 @@
+package com.example.newspeed.dto.user;
+
+public class UsersResponseDto {
+
+
+}
