+++ src/main/java/com/example/newspeed/service/UsersServiceImpl.java
@@ -1,11 +1,10 @@
 package com.example.newspeed.service;
 
-import com.example.newspeed.dto.*;
+import com.example.newspeed.dto.user.*;
 import com.example.newspeed.entity.Users;
 import com.example.newspeed.repository.UsersRepository;
 import com.example.newspeed.util.PasswordEncoder;
 import lombok.AllArgsConstructor;
-import org.apache.coyote.Response;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
